requirements.txt
alembic==1.8.1
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
bcrypt==3.2.2
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.3
Deprecated==1.2.13
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
greenlet==1.1.2
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.9.0
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.40
toml==0.10.2
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
